% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_resample_levels}
\alias{check_resample_levels}
\title{Check if resample has any errors}
\usage{
check_resample_levels(resample)
}
\arguments{
\item{resample}{A \code{rset} object}
}
\description{
Check if resample has any errors
}
