% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CV}
\alias{CV}
\alias{new_cv}
\title{CV Class}
\usage{
new_cv(options = NULL, keys = NULL, fit_param = NULL)
}
\arguments{
\item{options}{\code{\link{Options}} class object.}

\item{keys}{A named character vector of keys.}

\item{fit_param}{A list of fit parameters.}
}
\value{
A new \code{CV} class object
}
\description{
A class defines a cross-validation task specification.
}
\section{Super class}{
\code{\link[ml4e:Task]{ml4e::Task}} -> \code{CV}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{fit_param}}{A list of fit parameters.}

\item{\code{fit_call}}{A fit \code{call}.}

\item{\code{cv_pred}}{A \code{data.frame} of cross-validation predictions.}

\item{\code{test_pred}}{A \code{data.frame} of test predictions.}

\item{\code{pred}}{A \code{data.frame} of all predictions.}

\item{\code{cv_score}}{A list of cross-validation score.}

\item{\code{test_score}}{A list of test score.}

\item{\code{error_count}}{A list of error counts against folds}

\item{\code{error_count_str}}{A character of error counts.}

\item{\code{done}}{A logical if finished}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CV$new()}}
\item \href{#method-print}{\code{CV$print()}}
\item \href{#method-get_resample}{\code{CV$get_resample()}}
\item \href{#method-convert_data}{\code{CV$convert_data()}}
\item \href{#method-fit}{\code{CV$fit()}}
\item \href{#method-predict}{\code{CV$predict()}}
\item \href{#method-get_fitted}{\code{CV$get_fitted()}}
\item \href{#method-do}{\code{CV$do()}}
\item \href{#method-clone}{\code{CV$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ml4e" data-topic="Task" data-id="eval_call_args">}\href{../../ml4e/html/Task.html#method-eval_call_args}{\code{ml4e::Task$eval_call_args()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ml4e" data-topic="Task" data-id="print_task">}\href{../../ml4e/html/Task.html#method-print_task}{\code{ml4e::Task$print_task()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new \code{CV} class object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$new(options = NULL, keys = NULL, fit_param = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{options}}{A \link{Options} class object.}

\item{\code{keys}}{A named character vector of keys.}

\item{\code{fit_param}}{A list of fit parameters. If \code{NULL}, \code{keys} must have
"fit_param" key.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{CV} class object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_resample"></a>}}
\if{latex}{\out{\hypertarget{method-get_resample}{}}}
\subsection{Method \code{get_resample()}}{
Get a resample.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$get_resample(seed = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{A integer scalar seed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A resample of \code{rset} class object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-convert_data"></a>}}
\if{latex}{\out{\hypertarget{method-convert_data}{}}}
\subsection{Method \code{convert_data()}}{
Convert data by \code{self$data_call}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$convert_data(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A \code{data.frame} or \code{rsplit} object to be converted.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A converted data appropreate for the model input.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit model by \code{self$fit_call}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$fit(data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A \code{data.frame} or \code{rsplit} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A model fitted object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict by a model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$predict(fitted = NULL, new_data = NULL, ids = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fitted}}{A model fitted object.}

\item{\code{new_data}}{A \code{data.frame} to be predicted.}

\item{\code{ids}}{A integer vector of row indexes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} of predicted values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_fitted"></a>}}
\if{latex}{\out{\hypertarget{method-get_fitted}{}}}
\subsection{Method \code{get_fitted()}}{
Get a fitted object by all training data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$get_fitted()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-do"></a>}}
\if{latex}{\out{\hypertarget{method-do}{}}}
\subsection{Method \code{do()}}{
Run a cross-validation. The result will be stored in \code{self$result}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$do(cv_seed = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cv_seed}}{A logical scalar whether to use \code{self$cv_seed}. If
\code{FALSE}, then \code{self$search_seed} is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
