% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\name{data_catboost}
\alias{data_catboost}
\alias{data_catboost.data.frame}
\alias{data_catboost.rsplit}
\title{Catboost data function}
\usage{
data_catboost(object, ...)

\method{data_catboost}{data.frame}(object, formula, ...)

\method{data_catboost}{rsplit}(object, formula, ...)
}
\arguments{
\item{object}{A \code{data.frame} or \code{rset} object.}

\item{...}{Additinal arguments passed to the other S3 methods.}

\item{formula}{A \code{formula} object.}
}
\description{
Data function for catboost engine.
}
