% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\name{data_identity}
\alias{data_identity}
\alias{data_identity.data.frame}
\alias{data_identity.rsplit}
\title{Identity data function}
\usage{
data_identity(object, ...)

\method{data_identity}{data.frame}(object, ...)

\method{data_identity}{rsplit}(object, ...)
}
\arguments{
\item{object}{A \code{data.frame} or \code{rset} object.}

\item{...}{Not currently used.}
}
\description{
Data function for models that consumes \code{data.frame} as is.
}
