% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\name{data_lightgbm}
\alias{data_lightgbm}
\alias{data_lightgbm.data.frame}
\alias{data_lightgbm.rsplit}
\title{Lightgbm data function}
\usage{
data_lightgbm(object, ...)

\method{data_lightgbm}{data.frame}(object, formula, ...)

\method{data_lightgbm}{rsplit}(object, formula, ...)
}
\arguments{
\item{object}{A \code{data.frame} or \code{rset} object.}

\item{...}{Additinal arguments passed to the other S3 methods.}

\item{formula}{A \code{formula} object.}
}
\description{
Data function for lightgbm engine.
}
