% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\name{data_xgboost}
\alias{data_xgboost}
\alias{data_xgboost.data.frame}
\alias{data_xgboost.rsplit}
\title{Xgboost data function}
\usage{
data_xgboost(object, ...)

\method{data_xgboost}{data.frame}(object, formula, ...)

\method{data_xgboost}{rsplit}(object, formula, ...)
}
\arguments{
\item{object}{A \code{data.frame} or \code{rset} object.}

\item{...}{Additinal arguments passed to the other S3 methods.}

\item{formula}{A \code{formula} object.}
}
\description{
Data function for xgboost engine.
}
