% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\name{data_xy}
\alias{data_xy}
\alias{data_xy.data.frame}
\alias{data_xy.rsplit}
\title{XY data function}
\usage{
data_xy(object, ...)

\method{data_xy}{data.frame}(object, formula, ...)

\method{data_xy}{rsplit}(object, formula, ...)
}
\arguments{
\item{object}{A \code{data.frame} or \code{rset} object.}

\item{...}{Not currently used.}

\item{formula}{A \code{formula} object.}
}
\description{
Data function for models that consumes x (features) matrix and y (labels)
vector.
}
