% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{family_from_task}
\alias{family_from_task}
\title{Get family from task type}
\usage{
family_from_task(task_type)
}
\arguments{
\item{task_type}{A character scalar of task_type.}
}
\value{
A character scalar of family.
}
\description{
Get family from task type
}
\examples{
family_from_task("multiclass") # "multinomial"
family_from_task("binary") # "binomial"
family_from_task("regression") # "gaussian"

}
