% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_catboost.R
\name{fit_catboost_es}
\alias{fit_catboost_es}
\title{Fit GBDT by catboost package with early stopping}
\usage{
fit_catboost_es(data, eval, task_type, ...)
}
\arguments{
\item{data}{A training data of \code{catboost.Pool} class object.}

\item{eval}{A evaluation data of \code{catboost.Pool} class object.
Used for early stopping.}

\item{task_type}{A character scalar of task_type.}

\item{...}{Additional arguments passed to \link[catboost:catboost.train]{catboost::catboost.train}.}
}
\value{
A fitted object.
}
\description{
Fit GBDT by catboost package with early stopping
}
