% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_extraTrees.R
\name{fit_extraTrees}
\alias{fit_extraTrees}
\title{Fit extraTrees package with fraction mtry}
\usage{
fit_extraTrees(x, y, mtry = 0.2, ...)
}
\arguments{
\item{x}{A matrix of features.}

\item{y}{A vector of labels.}

\item{mtry}{A double scalar of \code{mtry} fraction.}

\item{...}{Additional arguments passed to \link[extraTrees:extraTrees]{extraTrees::extraTrees}.}
}
\value{
A fitted object.
}
\description{
Fit extraTrees package with fraction mtry
}
