% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_glmnet.R
\name{fit_glmnet_cv}
\alias{fit_glmnet_cv}
\title{Fit glmnet package with cross-validation}
\usage{
fit_glmnet_cv(x, y, family, use_min = TRUE, ...)
}
\arguments{
\item{x}{A matrix of features.}

\item{y}{A vector of labels.}

\item{family}{A character scalar of family.}

\item{use_min}{A logical scalar or binary integer wheather to use
\code{lambda.min}. If \code{FALSE}, \code{lambda.1se} is used.}

\item{...}{Additional arguments passed to \link[glmnet:cv.glmnet]{glmnet::cv.glmnet} and
\link[glmnet:glmnet]{glmnet::glmnet}.}
}
\value{
A fitted object.
}
\description{
Fit glmnet package with cross-validation
}
