% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_keras.R
\name{fit_keras_es}
\alias{fit_keras_es}
\title{Fit neural network by keras package with early stopping}
\usage{
fit_keras_es(
  x,
  y,
  eval_x,
  eval_y,
  task_type,
  keras_model,
  epoch = 10000L,
  patience = 20L,
  seed,
  ...
)
}
\arguments{
\item{x}{A matrix of features.}

\item{y}{A vector of labels.}

\item{eval_x}{A matrix of features for evaluation.}

\item{eval_y}{A vector of labels for evaluation.}

\item{task_type}{A character scalar of task_type.}

\item{keras_model}{A keras model object.}

\item{epoch}{A integer scalar of epoch.}

\item{patience}{A integer scalar of patience.}

\item{seed}{A integer scalar of random seed.}

\item{...}{Additional arguments passed to \link[keras:fit]{keras::fit}.}
}
\value{
A fitted object.
}
\description{
Fit neural network by keras package with early stopping
}
