% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_kernlab.R
\name{fit_kernlab}
\alias{fit_kernlab}
\title{Fit kernlab package with kernel parameters}
\usage{
fit_kernlab(x, data, kernel = "rbfdot", prob_model = FALSE, ...)
}
\arguments{
\item{x}{A formula.}

\item{data}{A data.frame.}

\item{kernel}{A character scalar of \code{kernel}.}

\item{prob_model}{A logical scalar.}

\item{...}{Additional arguments passed to \link[kernlab:ksvm]{kernlab::ksvm}.}
}
\value{
A fitted object.
}
\description{
Fit kernlab package with kernel parameters
}
