% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_lightgbm.R
\name{fit_lightgbm_es}
\alias{fit_lightgbm_es}
\title{Fit GBDT by lightgbm package with early stopping}
\usage{
fit_lightgbm_es(data, eval, task_type, ...)
}
\arguments{
\item{data}{A training data of \link[lightgbm:lgb.Dataset]{lightgbm::lgb.Dataset} class object.}

\item{eval}{A evaluation data of \link[lightgbm:lgb.Dataset]{lightgbm::lgb.Dataset} class object. Used
for early stopping.}

\item{task_type}{A character scalar of task_type.}

\item{...}{Additional arguments passed to \link[lightgbm:lgb.train]{lightgbm::lgb.train}].}
}
\value{
A fitted object.
}
\description{
Fit GBDT by lightgbm package with early stopping
}
