% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_ranger.R
\name{fit_ranger}
\alias{fit_ranger}
\title{Fit ranger package with fraction mtry}
\usage{
fit_ranger(formula, data, mtry = 0.2, ...)
}
\arguments{
\item{formula}{A \code{formula}.}

\item{data}{A \code{data.frame}.}

\item{mtry}{A double scalar of \code{mtry} fraction.}

\item{...}{Additional arguments passed to \link[ranger:ranger]{ranger::ranger}.}
}
\value{
A fitted object.
}
\description{
Fit ranger package with fraction mtry
}
