% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_RGF.R
\name{fit_RGF}
\alias{fit_RGF}
\title{Fit rgf by RGF package}
\usage{
fit_RGF(x, y, fast = FALSE, ...)
}
\arguments{
\item{x}{A matrix of features.}

\item{y}{A vector of labels.}

\item{fast}{A logical scalar wheather to use \code{RGF::FastRGF}.}

\item{...}{Additional arguments passed to \link[RGF:RGF_Classifier]{RGF::RGF_Classifier} or
\link[RGF:RGF_Regressor]{RGF::RGF_Regressor}.}
}
\value{
A fitted object.
}
\description{
Fit rgf by RGF package
}
