% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_xgboost.R
\name{fit_xgboost_es}
\alias{fit_xgboost_es}
\title{FitParamSpecs Class Constructor for xgboost Engine}
\usage{
fit_xgboost_es(data, eval, task_type, ...)
}
\arguments{
\item{data}{A training data of \code{\link[xgboost]{xgb.DMatrix}} class
object.}

\item{eval}{A evaluation data of \code{\link[xgboost]{xgb.DMatrix}} class
object. Used for early stopping.}

\item{task_type}{A character scalar of task_type.}

\item{...}{Additional arguments passed to \code{\link[xgboost]{xgb.train}}.}
}
\value{
A FitParamSpecs class object.

Fit GBDT by xgboost package with early stopping

A fitted object.
}
\description{
FitParamSpecs Class Constructor for xgboost Engine
}
