% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitParamSpec.R
\name{FitParamSpec}
\alias{FitParamSpec}
\alias{new_fit_param_spec}
\title{FitParamSpec Class}
\usage{
new_fit_param_spec(name = NULL, values = NULL, default = NULL, trans = NULL)
}
\arguments{
\item{name}{A character scalar of parameter name.}

\item{values}{A double, integer, character or logical vector of parameter
values.}

\item{default}{A default value. Must be a value within values.}

\item{trans}{A trans class object from scales package.}
}
\value{
A new \code{FitParamSpec} class object

A new \code{FitParamSpec} class object.
}
\description{
A class defines the model fit parameter specification.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{A parameter name.}

\item{\code{values}}{parameter values.}

\item{\code{default}}{A default parameter value.}

\item{\code{trans}}{A trans class object from sclaes package.}

\item{\code{nlevels}}{A lenght of parameters.}

\item{\code{type}}{A parameter type.}

\item{\code{range}}{A parameter range.}

\item{\code{has_default}}{A logical scalar wheather to have default value}

\item{\code{trans_values}}{Translated values}

\item{\code{trans_default}}{A translated default value}

\item{\code{dials_param}}{A param object from dials package}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FitParamSpec$new()}}
\item \href{#method-print}{\code{FitParamSpec$print()}}
\item \href{#method-get_dials_param}{\code{FitParamSpec$get_dials_param()}}
\item \href{#method-clone}{\code{FitParamSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new \code{FitParamSpec} class object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitParamSpec$new(name = NULL, values = NULL, default = NULL, trans = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character scalar of parameter name.}

\item{\code{values}}{A double, integer, character or logical vector of parameter
values.}

\item{\code{default}}{A default value. Must be a value within values.}

\item{\code{trans}}{A \code{trans} class object from \link{scales} package.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{FitParamSpec} class object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitParamSpec$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dials_param"></a>}}
\if{latex}{\out{\hypertarget{method-get_dials_param}{}}}
\subsection{Method \code{get_dials_param()}}{
Print object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitParamSpec$get_dials_param(quant_values = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quant_values}}{A logical scalar wheather to set fixed values.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \link{dials} class object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitParamSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
