% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure.R
\name{Measure}
\alias{Measure}
\alias{new_measure}
\title{Measure Class}
\usage{
new_measure(task_type = NULL, metrics = NULL)
}
\arguments{
\item{task_type}{A character scalar of task type. See
\link{supported_tasks} for tasks supported by a model/engine.}

\item{metrics}{A character vector of model metrics. Metrics must be
metrics of \link{supported_metrics}. If the vector is named, the names will be
used insted of metrics' name itself. If \code{NULL}, default metrics will be
set based on a \code{task_type}.}
}
\value{
A new \code{Measure} class object
}
\description{
A class defines metrics specification.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{task_type}}{A task type.}

\item{\code{metrics}}{A character vector of metrics.}

\item{\code{metric_names}}{A character vector of metric names.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Measure$new()}}
\item \href{#method-print}{\code{Measure$print()}}
\item \href{#method-do}{\code{Measure$do()}}
\item \href{#method-assert_metric_name}{\code{Measure$assert_metric_name()}}
\item \href{#method-is_minimize_metric}{\code{Measure$is_minimize_metric()}}
\item \href{#method-clone}{\code{Measure$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new \code{Measure} class object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$new(task_type = NULL, metrics = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task_type}}{A character scalar of task type. See
\link{supported_tasks} for tasks supported by a model/engine.}

\item{\code{metrics}}{A character vector of model metrics. Metrics must be
metrics of \link{supported_metrics}. If the vector is named, the names will be
used insted of metrics' name itself. If \code{NULL}, default metrics will be
set based on a \code{task_type}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Measure} class object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-do"></a>}}
\if{latex}{\out{\hypertarget{method-do}{}}}
\subsection{Method \code{do()}}{
Measure metrics of predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$do(pred = NULL, label_var = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pred}}{A \code{data.frame} of predictions.}

\item{\code{label_var}}{A character scalar of label variable name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{list} of metrics.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assert_metric_name"></a>}}
\if{latex}{\out{\hypertarget{method-assert_metric_name}{}}}
\subsection{Method \code{assert_metric_name()}}{
Assert a metric name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$assert_metric_name(metric_name = NULL, conversion = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metric_name}}{A character scalar of metric name.}

\item{\code{conversion}}{A logical scalar wheather to convert metric name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character scalar of metric name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_minimize_metric"></a>}}
\if{latex}{\out{\hypertarget{method-is_minimize_metric}{}}}
\subsection{Method \code{is_minimize_metric()}}{
Wheather the metric is minimize function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$is_minimize_metric(metric_name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metric_name}}{A character scalar of metric name. If NULL, the first
metric_name in \code{options$measure$metric_names} is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical scalar.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
