% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{Models}
\alias{Models}
\alias{new_models}
\title{Models Class}
\usage{
new_models(...)
}
\arguments{
\item{...}{Key and value pairs.}
}
\value{
A new \code{Models} class object.
}
\description{
Key-value dictionary contains \code{\link{Model}} class objects.
This class is passed to \code{\link{Options}} class.
}
\section{Super classes}{
\code{\link[Dict:Dict]{Dict::Dict}} -> \code{\link[ml4e:Option]{ml4e::Option}} -> \code{Models}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Models$new()}}
\item \href{#method-print}{\code{Models$print()}}
\item \href{#method-add}{\code{Models$add()}}
\item \href{#method-clone}{\code{Models$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="Dict" data-topic="Dict" data-id="get">}\href{../../Dict/html/Dict.html#method-get}{\code{Dict::Dict$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Dict" data-topic="Dict" data-id="has">}\href{../../Dict/html/Dict.html#method-has}{\code{Dict::Dict$has()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Dict" data-topic="Dict" data-id="sort">}\href{../../Dict/html/Dict.html#method-sort}{\code{Dict::Dict$sort()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ml4e" data-topic="Option" data-id="clear">}\href{../../ml4e/html/Option.html#method-clear}{\code{ml4e::Option$clear()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ml4e" data-topic="Option" data-id="print_items">}\href{../../ml4e/html/Option.html#method-print_items}{\code{ml4e::Option$print_items()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ml4e" data-topic="Option" data-id="remove">}\href{../../ml4e/html/Option.html#method-remove}{\code{ml4e::Option$remove()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new \code{Models} class object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Models$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Key and value pairs.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Models} class object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Models$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
Adds key-value objects to the dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Models$add(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Key and value pairs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Models$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
