% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_extraTrees.R
\name{new_fit_param_specs_extraTrees}
\alias{new_fit_param_specs_extraTrees}
\title{FitParamSpecs Class Constructor for extraTrees Engine}
\usage{
new_fit_param_specs_extraTrees(task_type = NULL)
}
\arguments{
\item{task_type}{A character scalar of task type.}
}
\value{
A FitParamSpecs class object.
}
\description{
FitParamSpecs Class Constructor for extraTrees Engine
}
