% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_glmnet.R
\name{new_fit_param_specs_glmnet_cv}
\alias{new_fit_param_specs_glmnet_cv}
\title{FitParamSpecs Class Constructor for glmnet_cv Engine}
\usage{
new_fit_param_specs_glmnet_cv()
}
\value{
A FitParamSpecs class object.
}
\description{
FitParamSpecs Class Constructor for glmnet_cv Engine
}
