% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_kernlab.R
\name{new_fit_param_specs_kernlab}
\alias{new_fit_param_specs_kernlab}
\title{FitParamSpecs Class Constructor for kernlab Engine}
\usage{
new_fit_param_specs_kernlab(kernel = NULL, type = NULL)
}
\arguments{
\item{kernel}{A character scalar of \code{kernel}.}

\item{type}{A character scalar of \code{type}.}
}
\value{
A FitParamSpecs class object.
}
\description{
FitParamSpecs Class Constructor for kernlab Engine
}
