% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_kernlab.R
\name{new_fit_param_specs_kernlab_rbf}
\alias{new_fit_param_specs_kernlab_rbf}
\title{FitParamSpecs Class Constructor for kernlab Engine with RBF Kernel}
\usage{
new_fit_param_specs_kernlab_rbf()
}
\value{
A FitParamSpecs class object.
}
\description{
FitParamSpecs Class Constructor for kernlab Engine with RBF Kernel
}
