% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_kknn.R
\name{new_fit_param_specs_kknn}
\alias{new_fit_param_specs_kknn}
\title{FitParamSpecs Class Constructor for kknn Engine}
\usage{
new_fit_param_specs_kknn()
}
\value{
A FitParamSpecs class object.
}
\description{
FitParamSpecs Class Constructor for kknn Engine
}
