% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_lightgbm.R
\name{new_fit_param_specs_lightgbm_es}
\alias{new_fit_param_specs_lightgbm_es}
\title{FitParamSpecs Class Constructor for lightgbm Engine}
\usage{
new_fit_param_specs_lightgbm_es()
}
\value{
A FitParamSpecs class object.
}
\description{
FitParamSpecs Class Constructor for lightgbm Engine
}
