% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_ranger.R
\name{new_fit_param_specs_ranger}
\alias{new_fit_param_specs_ranger}
\title{FitParamSpecs Class Constructor for ranger Engine}
\usage{
new_fit_param_specs_ranger(task_type = NULL, splitrule = NULL)
}
\arguments{
\item{task_type}{A character scalar of task type.}

\item{splitrule}{A character scalar of splitrule.}
}
\value{
A FitParamSpecs class object.
}
\description{
FitParamSpecs Class Constructor for ranger Engine
}
