% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_rpart.R
\name{new_fit_param_specs_rpart}
\alias{new_fit_param_specs_rpart}
\title{FitParamSpecs Class Constructor for rpart Engine}
\usage{
new_fit_param_specs_rpart()
}
\value{
A FitParamSpecs class object.
}
\description{
FitParamSpecs Class Constructor for rpart Engine
}
