% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Options.R
\name{Options}
\alias{Options}
\alias{new_options}
\title{Options Class}
\usage{
new_options(
  formulas = NULL,
  datasets = NULL,
  seeds = NULL,
  models = NULL,
  measure = NULL,
  show_progress = TRUE,
  keep_data = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{formulas}{A \link{Formulas} class object.}

\item{datasets}{A \link{Datasets} class object.}

\item{seeds}{A \link{Seeds} class object.}

\item{models}{A \link{Models} class object.}

\item{measure}{A \link{Measure} class object. If \code{NULL}, a Measure class
object with default metrics are used.}

\item{show_progress}{A logical scalar wheather to show a progress bar.}

\item{keep_data}{A logical scalar wheather to keep original data.
If \code{TRUE}, the original dataset are kept in predicted values and
cross-validation folds.}

\item{parallel}{A logical scalar wheather to calculate parallely.}
}
\value{
A new \code{Options} class object
}
\description{
A class stores all machine learning project \link{Option}. This class checks if
the option have valid values each other. (e.g. All data in Datasets have the
label variable described in the formulas, or all models support the task
type, etc.)

Methods that do actual tasks are not defined in the classs. The concept is,
gather the all options in one place and pass it to \link{Task} and \link{Project}
classes.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{formulas}}{A \code{Formulas} class object.}

\item{\code{datasets}}{A \code{Datasets} class object.}

\item{\code{seeds}}{A \code{Seeds} class object.}

\item{\code{models}}{A \code{Models} class object.}

\item{\code{measure}}{A \code{Measure} class object.}

\item{\code{show_progress}}{Wheather to show progress messages.}

\item{\code{keep_data}}{Wheather to keep data.}

\item{\code{parallel}}{Wheather to calculate parallely.}

\item{\code{labels}}{A vector of labels (response variable).}

\item{\code{label_var}}{A label (response) variable name.}

\item{\code{label_levels}}{Levels of label (response) variable.}

\item{\code{task_type}}{A task type.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Options$new()}}
\item \href{#method-print}{\code{Options$print()}}
\item \href{#method-print_options}{\code{Options$print_options()}}
\item \href{#method-clone}{\code{Options$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new \code{Options} class object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$new(
  formulas = NULL,
  datasets = NULL,
  seeds = NULL,
  models = NULL,
  measure = NULL,
  show_progress = TRUE,
  keep_data = FALSE,
  parallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formulas}}{A \link{Formulas} class object.}

\item{\code{datasets}}{A \link{Datasets} class object.}

\item{\code{seeds}}{A \link{Seeds} class object.}

\item{\code{models}}{A \link{Models} class object.}

\item{\code{measure}}{A \link{Measure} class object. If \code{NULL}, a Measure class
object with default metrics are used.}

\item{\code{show_progress}}{A logical scalar wheather to show a progress bar.}

\item{\code{keep_data}}{A logical scalar wheather to keep original data.
If \code{TRUE}, the original dataset are kept in predicted values and
cross-validation folds.}

\item{\code{parallel}}{A logical scalar wheather to calculate parallely.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Options} class object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_options"></a>}}
\if{latex}{\out{\hypertarget{method-print_options}{}}}
\subsection{Method \code{print_options()}}{
Print Options object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$print_options(title = " Options ", key_width = 10L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{A character scalar of printed title.}

\item{\code{key_width}}{A integer scalar of key column width.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Options$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
