% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_RGF.R
\name{predict_RGF}
\alias{predict_RGF}
\title{Predict rgf by RGF package}
\usage{
predict_RGF(object, x)
}
\arguments{
\item{object}{A fitted object.}

\item{x}{A matrix of features.}
}
\value{
Predicted values.
}
\description{
Predict rgf by RGF package
}
