% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocCalls.R
\name{preproc_identity}
\alias{preproc_identity}
\title{Preproc call that does nothing}
\usage{
preproc_identity(data)
}
\arguments{
\item{data}{A \code{data.frame} object.}
}
\value{
Input \code{data.frame}
}
\description{
Preproc call that does nothing
}
