% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%<>\%}
\alias{.data}
\alias{predict.glmnet}
\alias{contr.dummy}
\alias{contr.ordinal}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{glmnet}{\code{\link[glmnet]{predict.glmnet}}}

  \item{kknn}{\code{\link[kknn]{contr.dummy}}, \code{\link[kknn]{contr.ordinal}}}

  \item{magrittr}{\code{\link[magrittr]{\%<>\%}}, \code{\link[magrittr]{\%>\%}}}

  \item{rlang}{\code{\link[rlang]{.data}}}
}}

