% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchBase.R
\name{SearchBase}
\alias{SearchBase}
\title{SearchBase Class}
\description{
A base class describe a hyper parameter search task.
}
\section{Super class}{
\code{\link[ml4e:Task]{ml4e::Task}} -> \code{SearchBase}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{result}}{A task result.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SearchBase$new()}}
\item \href{#method-print}{\code{SearchBase$print()}}
\item \href{#method-get_rank}{\code{SearchBase$get_rank()}}
\item \href{#method-get_best_param}{\code{SearchBase$get_best_param()}}
\item \href{#method-clone}{\code{SearchBase$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ml4e" data-topic="Task" data-id="eval_call_args">}\href{../../ml4e/html/Task.html#method-eval_call_args}{\code{ml4e::Task$eval_call_args()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ml4e" data-topic="Task" data-id="print_task">}\href{../../ml4e/html/Task.html#method-print_task}{\code{ml4e::Task$print_task()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new \code{SearchBase} class object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchBase$new(options = NULL, keys = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{options}}{A \link{Options} class object.}

\item{\code{keys}}{A named character vector of keys.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{SearchBase} class object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchBase$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_rank"></a>}}
\if{latex}{\out{\hypertarget{method-get_rank}{}}}
\subsection{Method \code{get_rank()}}{
Get parameters ranking tables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchBase$get_rank(metric_name = NULL, n = 5L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metric_name}}{A character scalar of metric name. If \code{NULL}, the
first metric_name in \code{self$options$measure$metric_names} is used.}

\item{\code{n}}{A integer scalar of a number of the ranking.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of parameters ranking by keys
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_best_param"></a>}}
\if{latex}{\out{\hypertarget{method-get_best_param}{}}}
\subsection{Method \code{get_best_param()}}{
Get best parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchBase$get_best_param(metric_name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metric_name}}{A character scalar of metric name. If \code{NULL}, the
first metric_name in \code{self$options$measure$metric_names} is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of parameters.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
