% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{supported_engines}
\alias{supported_engines}
\alias{supported_tasks}
\alias{supported_metrics}
\title{Supported specifications by ml4e package}
\usage{
supported_engines()

supported_tasks(engine = NULL)

supported_metrics()
}
\arguments{
\item{engine}{A character scalar of model engine.}
}
\description{
Supported specifications by ml4e package
}
\examples{
supported_engines()

supported_tasks("glm")

supported_metrics()

}
